server/requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.0
beanie>=1.23.6
python-jose[cryptography]>=3.3.0
argon2-cffi>=21.3.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx
pydantic-settings
pydantic
pydantic[email]
python-multipart
pydantic-core
pymongo